added optional htpasswd protecion

+++ startup-mediawiki.sh
@@ -1,13 +1,26 @@
 #/bin/bash
 
-chmod a+r /LocalSettings.php
-sed -i 's/^$wgServer/#$wgServer/g' /LocalSettings.php
-ln -s /LocalSettings.php /usr/share/nginx/html/LocalSettings.php
+if [ -f /LocalSettings.php ]
+then
+  echo ">> enabling LocalSettings"
+  chmod a+r /LocalSettings.php
+  sed -i 's/^$wgServer/#$wgServer/g' /LocalSettings.php
+  ln -s /LocalSettings.php /usr/share/nginx/html/LocalSettings.php
+fi
+
+if [ -f /htpasswd ]
+then
+  echo ">> enabling htpasswd protection"
+  chmod a+r /htpasswd
+  sed -i 's/#auth_basic/auth_basic/g' /htpasswd
+  ln -s /htpasswd /etc/nginx/htpasswd
+fi
 
 chmod a+rw -R /usr/share/nginx/data/
 
 if [ -d /uploads ]
 then
+  echo ">> enabling uploads volume"
   mv /usr/share/nginx/html/images/* /uploads/
   rm -f /usr/share/nginx/html/images
   chmod a+rwx -R /uploads
